routes /authors /collections are now complete and functional. Need to fix an error on PUT /illustrator/:id

+++ routes/authors.js
@@ -68,10 +68,10 @@ const authors = {
     const db = database.get()
 
     if (typeof request.body.name === 'undefined') {
-      return response.status(422).json({
-        status: 422,
+      return response.status(412).json({
+        status: 412,
         success: false,
-        message: 'Unprocessable entity'
+        message: 'Body missing name field'
       })
     }
 
@@ -111,10 +111,21 @@ const authors = {
     const ObjectId = require('mongodb').ObjectId
 
     try {
+      if (request.body.id === undefined) {
+        return response.status(412).json({
+          status: 412,
+          success: false,
+          message: 'Body missing id field'
+        })
+      }
+
       const author = await db.collection('authors').findOne(
         {
           _id: ObjectId(request.params.id),
           userId: request.decoded.userId
+        },
+        {
+          name: 1
         })
 
       if (!author) {
@@ -132,7 +143,18 @@ const authors = {
         },
         {
           $push: {
-            books: {id: request.body.name}
+            books: {id: request.body.id}
+          }
+        })
+
+      await db.collection('books').update(
+        {
+          _id: ObjectId(request.body.id),
+          userId: request.decoded.userId
+        },
+        {
+          $set:{
+            author: author.name
           }
         })
 
@@ -160,6 +182,14 @@ const authors = {
           name: 1
         })
 
+      if (!target) {
+        return response.status(404).json({
+          status: 404,
+          success: false,
+          message: 'Author not found'
+        })
+      }
+
       const doc = await db.collection('authors').findOneAndDelete(
         {
           _id: ObjectId(request.params.id),
@@ -174,8 +204,11 @@ const authors = {
           },
           {
             $set: {
-              collection: ''
+              author: ''
             }
+          },
+          {
+            multi: true
           })
 
         return response.status(200).json({
